Handle EROFS errors
A EROFS will be raised when trying to write a dir onto a read-only
filesystem.  However, if the dir already is there, then it should be
treated like an EEXIST (since EROFS can be raised by trying to create a
dir over a mount point, where the mount point is not read-only, but you
still can't just clobber over it).

If the dir doesn't already exist, then the EROFS will be accurately
reported as the reason why it could not be created.

(Copied from 17ee84f2073a262a65e201acedb71e0d20bf9273)